Convert level and pattern difficulty fields from int to float
- Changed level field to float in ChartDataNew.cs and ChartData.cs
- Changed all 8 pattern difficulty fields to float in PatternDifficulty.cs
- Updated PlayRecord.cs pattern fields to float
- Added :F1 formatting in CustomChartWriter.cs for 1 decimal precision
- Updated CustomChartParser.cs to parse floats with InvariantCulture
- Updated PlayerProfile.cs GetPatternValue() to return float

All fields now support 1 decimal place precision (e.g., 15.5, 17.3)

+++ Synth/Assets/Play/ChartData.cs
@@ -27,7 +27,7 @@ public class ChartData
     [Header("난이도 정보")]
     public string difficulty = "Normal"; // Easy, Normal, Hard, Expert, Master, Special
     public int keyCount = 4; // 4K, 5K, 6K, 7K, 8K, 10K
-    public int level = 1; // 난이도 레벨 (1-20)
+    public float level = 1.0f; // 난이도 레벨 (1-20, 소수점 1자리)
 
     [Header("차트 제작 정보 (선택)")]
     public string chartAuthor = ""; // 차트 제작자
@@ -118,7 +118,7 @@ public class ChartData
         // 난이도 정보
         difficulty = "Normal";
         keyCount = 4;
-        level = 1;
+        level = 1.0f;
 
         // 차트 제작 정보
         chartAuthor = "";
